Update RAG using Snowflake.sql
the 'request' method which is not longer valid and resulting in errors since Python UDFs do not have direct access to external networks, including S3. This restriction is preventing them from downloading or reading files directly from an S3 bucket, even when the stage is created successfully. 
Instead SnowflakeFile APIs can be directly used to process the file in your Python UDF.